fix(webauthn): expose WebAuthnSubject type + Request augmentation from package barrel
The Express Request.webauthnSubject augmentation was declared inside an
internal route file with no barrel export. Consumers writing upstream
auth middleware to set req.webauthnSubject (per README guidance) lacked
both the type and the type augmentation.

Moves WebAuthnSubject + the declare-global block to a dedicated
src/request.mts file exported from the barrel. The internal
registrationOptions.mts re-exports the type for backward compat.
Augmentation now reliably merges in consumer TypeScript builds.

Cross-refs: Claude Round 1 C2
